Moved constant values to constants file
