+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "annapurna-bhavan-website",
-  "version": "0.14.0",
+  "version": "0.14.1",
   "private": true,
   "scripts": {
     "dev": "next dev --turbopack",
